refactor: mapper 함수를 dto 파일에서 분리
- dto의 의미를 보존하기 위해서 별도 mapper 디렉토리를 만들어서 추가함
